+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessJSFilesTask.java
@@ -109,6 +109,7 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
                     closureConfig.getCompilationLevel().setOptionsForCompilationLevel(options);
                     options.setOutputCharset(charset);
                     options.setLanguageIn(closureConfig.getLanguage());
+                    options.setAngularPass(closureConfig.getAngularPass());
 
                     File sourceMapResult = new File(minifiedFile.getPath() + ".map");
                     if (closureConfig.getSourceMapFormat() != null) {